refactor: change the data type of 'filesize' field to int64
The previous data type of 'filesize' field in the model was string, which was not the most appropriate data type to represent the file size. To improve the efficiency of data storage and query, we have changed the data type of 'filesize' field to int64.

Using int64 data type for 'filesize' field has multiple advantages. Firstly, it has fixed length, which makes data storage and query faster. Secondly, it is a standard data type for representing integers, which makes it easier for developers to work with.

Therefore, this commit refactors the 'filesize' field from string to int64 data type.